% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number.R
\name{copy_number}
\alias{copy_number}
\title{Get copy number.}
\usage{
copy_number(x, log2_transform = FALSE, offset = 1)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}}.}

\item{log2_transform}{A logical scalar. Default to FALSE.}

\item{offset}{A numeric scalar. It is only applicable when log2 transform
is carried out. Default to 1.}
}
\value{
A matrix.
}
\description{
The function returns copy number.
}
\note{
\code{\link[minfi]{getCN}} returns the log2 value of meth + unmeth. If
  the sum is zero, \code{NA} is returned.
}
