% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-repo.R
\name{create_idat_table}
\alias{create_idat_table}
\title{Create IDAT table}
\usage{
create_idat_table(base, rda = NULL)
}
\arguments{
\item{base}{A character scalar of base directory path which has
IDAT files.}

\item{rda}{A character scalar of file path which saves the return
value. Default is NULL, - do not save.}
}
\value{
An \code{IdatTable} object.
}
\description{
Create IDAT table from base directory
}
