% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_bisulfite_converstion_type_1_green_qc}
\alias{get_bisulfite_converstion_type_1_green_qc}
\title{Get QC metric of bisulfite Conversion of type I probe design green probes.}
\usage{
get_bisulfite_converstion_type_1_green_qc(
  ctrl_probes,
  background_offset = 3000
)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}

\item{background_offset}{background correction offset. Default to 3000.}
}
\value{
A data frame of QC metrics.
}
\description{
There are conflictings between Illumina Infinium HD Methylation Assay
(Document # 15019519 v10) and \href{https://help.dragenarray.illumina.com/product-guides/output-files#methyl_qc_report}{DRAGEN documentation}.
The Illumina Infinium HD Methylation Assay said C1-C3 and U1-U3 are for
bisulfite conversion I green channel. But based on DRAGEN document, C3 and U3
are actually on red channel. I use probes C1-C2 and U1-U2 in this function.
}
