% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_bisulfite_converstion_type_1_red_qc}
\alias{get_bisulfite_converstion_type_1_red_qc}
\title{Get QC metric of bisulfite Conversion of type I probe design red probes.}
\usage{
get_bisulfite_converstion_type_1_red_qc(ctrl_probes, background_offset = 3000)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}

\item{background_offset}{background correction offset. Default to 3000.}
}
\value{
A data frame of QC metrics.
}
\description{
There are conflictings between Illumina Infinium HD Methylation Assay
(Document # 15019519 v10) and [DRAGEN documentation](https://help.dragenarray.illumina.com/product-guides/output-files#methyl_qc_report).
The Illumina Infinium HD Methylation Assay said C4-C6 and U4-U6 are for
bisulfite conversion I green channel. But based on DRAGEN document, C3 and U3
are actually on red channel too. I use probes C3-C6 and U3-U6 in this function.
}
