% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_bisulfite_converstion_type_2_qc}
\alias{get_bisulfite_converstion_type_2_qc}
\title{Get QC metric of bisulfite Conversion of type II probe design probes.}
\usage{
get_bisulfite_converstion_type_2_qc(ctrl_probes, background_offset = 3000)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}

\item{background_offset}{background correction offset. Default to 3000.}
}
\value{
A data frame of QC metrics.
}
\description{
Only on red channel.
}
