% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_control_probe_qc_metrics}
\alias{get_control_probe_qc_metrics}
\title{Get QC metrics of control probes.}
\usage{
get_control_probe_qc_metrics(rgset, background_offset = 3000)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{background_offset}{background correction offset. Default to 3000.}
}
\value{
A data frame of QC metrics.
}
\description{
Get QC metrics of control probes.
}
