% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-pvalue.R
\name{get_detection_p_value_qc_metrics}
\alias{get_detection_p_value_qc_metrics}
\title{Summarize detection p-values by sample.}
\usage{
get_detection_p_value_qc_metrics(rgset)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}
}
\value{
A \code{data.frame} of summarzing detection p-value for QC purpose.
}
\description{
Summarize detection p-values by sample.
}
