% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_gse}
\alias{get_gse}
\title{Create \code{RGChannelSet} object from GSE accession.}
\usage{
get_gse(gse_acc, data_dir)
}
\arguments{
\item{gse_acc}{A character scalar of a single GSE accession.}

\item{data_dir}{A character scalar of the directory stored IDAT files.}
}
\value{
A \code{\link[minfi]{RGChannelSet}} object.
}
\description{
Download IDAT files if available and then create
\code{\link[minfi]{RGChannelSet}}.
}
