% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_hybridization_green_qc}
\alias{get_hybridization_green_qc}
\title{Get QC metric of hybridization probe.}
\usage{
get_hybridization_green_qc(ctrl_probes)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}
}
\value{
A data frame of QC metrics.
}
\description{
It is only on green channel.
}
