% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-repo.R
\name{get_idat_file_list}
\alias{get_idat_file_list}
\title{Get IDAT file list}
\usage{
get_idat_file_list(base)
}
\arguments{
\item{base}{A character scalar of base directory path which has
IDAT files.}
}
\value{
A character vector of IDAT file paths.
}
\description{
Get IDAT file list
}
\details{
This function search the directory base (possibly including
subdirectories depending on the argument recursive for “sample sheet”
files (see below). These files are identified solely on the base of
their filename given by the arguments pattern "_(Grn|Red)\\.idat$"
and ignore case.
}
