% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_nonpolymorphic_green_qc}
\alias{get_nonpolymorphic_green_qc}
\title{Get QC metric of non-polymorphic green probes.}
\usage{
get_nonpolymorphic_green_qc(ctrl_probes)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}
}
\value{
A data frame of QC metrics.
}
\description{
There are five extra NP (G) probes of uncertain use. which I do not use
for non-polymorphic probe QC metric computation.
}
