% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_nonpolymorphic_red_qc}
\alias{get_nonpolymorphic_red_qc}
\title{Get QC metric of non-polymorphic red probes.}
\usage{
get_nonpolymorphic_red_qc(ctrl_probes)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}
}
\value{
A data frame of QC metrics.
}
\description{
Get QC metric of non-polymorphic red probes.
}
