% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_metrics.R
\name{get_qc_metrics}
\alias{get_qc_metrics}
\title{Get QC metrics.}
\usage{
get_qc_metrics(rgset, background_offset = 3000, verbose = TRUE)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{background_offset}{background correction offset. Default to 3000.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
An object of \code{\link[minfi]{RGChannelSet-class}} with QC metrics
  added to phenotype data.
}
\description{
Get QC metrics.
}
