% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_metrics.R
\name{get_sentrix_id}
\alias{get_sentrix_id}
\title{Get Sentrix IDs.}
\usage{
get_sentrix_id(sentrix_id_position)
}
\arguments{
\item{internal_sample_id}{internal sample ID}
}
\value{
a character vector of Sentrix ID
}
\description{
Get Sentrix IDs.
}
