% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_metrics.R
\name{get_sentrix_position}
\alias{get_sentrix_position}
\title{Get Sentrix position.}
\usage{
get_sentrix_position(sentrix_id_position)
}
\arguments{
\item{internal_sample_id}{internal sample ID}
}
\value{
a character vector of Sentrix position
}
\description{
Get Sentrix position.
}
