% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_specificity_type_1_green_qc}
\alias{get_specificity_type_1_green_qc}
\title{Get QC metric of specificity of type I probe design green probes.}
\usage{
get_specificity_type_1_green_qc(ctrl_probes)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}
}
\value{
A data frame of QC metrics.
}
\description{
Get QC metric of specificity of type I probe design green probes.
}
