% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{get_target_removal_qc}
\alias{get_target_removal_qc}
\title{Get QC metric of target removal probes.}
\usage{
get_target_removal_qc(ctrl_probes, background_offset = 3000)
}
\arguments{
\item{ctrl_probes}{The object returned by \code{\link{control_probe_intensities}}.}

\item{background_offset}{background correction offset. Default to 3000.}
}
\value{
A data frame of QC metrics.
}
\description{
It is only on green channel.
}
