% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsci_pal_d}
\alias{ggsci_pal_d}
\title{ggsci pal discrete.}
\usage{
ggsci_pal_d(ggsci_pal_name = "jco", pal_size = 5, ...)
}
\arguments{
\item{ggsci_pal_name}{ggsci palette name.}

\item{pal_size}{an integer of how many colors needed.}

\item{...}{other arguments passed to colorRampPalette or ggsci pal
functions (e.g. \code{\link[ggsci]{pal_jco}}).}
}
\value{
a character vector of colors.
}
\description{
ggsci pal discrete.
}
