% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_report}
\alias{init_report}
\title{Initiate report directory.}
\usage{
init_report(rgset, report_dir)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting directory.}
}
\value{
A character vector of sample directories.
}
\description{
Create the report directory and subdirectories of samples if they do not
exist. The names of sample directories are specified by \code{Sample_Name}
column of phenotype data returned by \code{\link[minfi]{pData}}.
}
