% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{is_annotated}
\alias{is_annotated}
\title{Is an object of minfi classes annotated?}
\usage{
is_annotated(x)
}
\arguments{
\item{x}{An object of minfi classes.}
}
\value{
A bool vector of length 1.
}
\description{
See minif classes in \href{https://bioconductor.org/packages/release/bioc/vignettes/minfi/inst/doc/minfi.html#2_minfi}{The minfi User's Guide}.
}
