% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize.dkfz}
\alias{normalize.dkfz}
\title{Normalization DKFZ.}
\usage{
normalize.dkfz(rgset, map_to_genome = TRUE)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{map_to_genome}{A logical scalar if an object of \code{\link[minfi]{MethylSet-class}}
or \code{\link[minfi]{GenomicMethylSet-class}} will be returned. Default
to TRUE.}
}
\value{
An object of \code{\link[minfi]{GenomicMethylSet-class}} or
  \code{\link[minfi]{MethylSet-class}}.
}
\description{
DKFZ implement the normalization method used in the paper
\href{https://www.nature.com/articles/nature26000}{Capper et al. DNA methylation-based classification of central nervous system tumours. Nature (2018)}.
}
