% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize.methylcnv}
\alias{normalize.methylcnv}
\title{Normalization MethylCNV.}
\usage{
normalize.methylcnv(rgset, target_value = 13, offset = 1)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{target_value}{A numeric value to which the median of log2 copy number
on each array is adjusted. Default to 13.}

\item{offset}{A numeric scalar. It is only applicable when log2 transform
is carried out. Default to 1.}
}
\value{
An object of \code{\link[minfi]{GenomicRatioSet-class}}.
}
\description{
Normalization method used in MethylCNV pipeline, which adjust the median of
log2 copy number on each array to a target value of 13.
}
\details{
The method was used in copy number variant pipeline described in
the paper Feng, G. A Statistical Method to Estimate DNA Copy Number from
Illumina High-Density Methylation Arrays. Systems Biomedicine (2013). It
adjust the median of log2 copy number on each array to a target value of 13.
}
