% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-value-distribution-plots.R
\name{plot_beta_value_histogram_fill_by_sample}
\alias{plot_beta_value_histogram_fill_by_sample}
\title{Beta value distribution histgram fill by sample.}
\usage{
plot_beta_value_histogram_fill_by_sample(x, output_file, ...)
}
\arguments{
\item{x}{An object of minfi classes.}

\item{output_file}{output plot file.}

\item{...}{arguments passed to \code{\link[ggplot2]{ggsave}}.}
}
\value{
A ggplot object of the plot.
}
\description{
Beta value distribution histgram fill by sample.
}
