% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{plot_control_probes}
\alias{plot_control_probes}
\title{Quality control plot of control probes in terms of raw intensities.}
\usage{
plot_control_probes(x, s = 1)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{s}{An integer scalar. Default to 1.}
}
\value{
A list of \code{df}, \code{red_plot}, \code{green_plot},
  \code{norm_plot}, \code{independent_control_plot}.
  The {df} is a \code{data.frame} containing the following columns:
  Address, Type, Color, ExtendedType, Channel, Expected_Grn, Expected_Red,
  Sample, and Intensity.
  Each of the \code{red_plot}, \code{green_plot}, \code{norm_plot} and
  \code{independent_control_plot} is an object of \code{\link[ggplot2]{ggplot}}
  class. They are plots of sample-dependent controls of red and green channel,
  normalization probes and sample-independent controls, respectively.
}
\description{
Quality control plot of control probes in terms of raw intensities.
}
