% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_control_probes}
\alias{qc_control_probes}
\title{Quality control}
\usage{
qc_control_probes(
  rgset,
  report_dir,
  plot_file = "control_probes.png",
  height = 10,
  width = 12,
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting directory.}

\item{plot_file}{A character scalar of the filename of control probe plot.
Default to "control_probes.png".}

\item{height}{A numeric scalar of control probe plot height. Default to 10.}

\item{width}{A numeric scalar of control probe plot width. Default to 12.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any arguments can be passed to \code{\link[ggplot2]{ggsave}}.}
}
\value{
A list of the return of \code{\link{plot_control_probes}} function
  (invisible).
}
\description{
Quality control are performed in terms of observed intensities of control
probes with expected value, detection p-values, and gender.
}
