% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pipe.R
\name{qc_pipe}
\alias{qc_pipe}
\title{QC pipeline.}
\usage{
qc_pipe(
  rgset,
  output,
  qc_csv = "qc.csv",
  qc_excel = "qc.xlsx",
  beta_value_distribution_file = "beta_value_histogram_fill_by_sample.png",
  background_offset = 3000,
  verbose = TRUE
)
}
\arguments{
\item{rgset}{An object of \code{\link[minfi]{RGChannelSet-class}}.}

\item{output}{output directory.}

\item{qc_csv}{QC metrics CSV file.}

\item{qc_excel}{QC metrics Excel file.}

\item{beta_value_distribution_file}{a plot file for beta value distribution.}

\item{background_offset}{background correction offset. Default to 3000.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
An object of \code{\link[minfi]{RGChannelSet-class}} with QC metrics
  added to phenotype data.
}
\description{
QC pipeline.
}
