% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_metharray_exp}
\alias{read_metharray_exp}
\title{Read methylation array with EPIC v2 support.}
\usage{
read_metharray_exp(...)
}
\arguments{
\item{...}{arguments passed onto \code{\link[minfi]{read.metharray.exp}}.}
}
\value{
x An object that has \code{sampleNames} method, for example,
  \code{\link[minfi]{RGChannelSet-class}}.
}
\description{
It is a wrapper of \code{\link[minfi]{read.metharray.exp}}, but with EPIC v2
support by setting \code{annotation} attribute.
}
