% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect.R
\name{remove_batch_effect}
\alias{remove_batch_effect}
\title{Remove batch effect.}
\usage{
remove_batch_effect(x, batch = NULL, batch2 = NULL, method = c("mum",
  "cn", "beta"), offset = 1, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}} if \code{method} is "mum".
An object of \code{\link[minfi]{RatioSet-class}} or
\code{\link[minfi]{GenomicRatioSet-class}} if \code{method} is "cn" or
"beta".}

\item{batch}{A character scalar of the column name of interested batch in
phenotype \code{data.frame}. Default to \code{NULL}.}

\item{batch2}{A character scalar of the column name of interested batch2 in
phenotype \code{data.frame}. Default to \code{NULL}.}

\item{method}{A character scalar of methods, including "mum" (methylation
and unmethylation separately), "cn" (copy number), "beta" (beta-value).
Default to "mum".}

\item{offset}{The offset is chosen to avoid log2 transform on zeros.
Default to 1. Only valid for "mum" method.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Other arguments can be passed to \code{\link[limma]{removeBatchEffect}}.}
}
\value{
An object of \code{\link[minfi]{GenomicMethylSet-class}}.
}
\description{
Remove batch effect.
}
\details{
Batch effect is removed by calling \code{\link[limma]{removeBatchEffect}}.
  The methylation and unmethylation signals are log2 transformed, removed
  batch effect, and transformed back.
}
