% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect.R
\name{remove_batch_effect_cn}
\alias{remove_batch_effect_cn}
\title{Remove batch effect: copy number}
\usage{
remove_batch_effect_cn(x, batch = NULL, batch2 = NULL, offset = 1, ...)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{RatioSet-class}} or
\code{\link[minfi]{GenomicRatioSet-class}}.}

\item{batch}{factor or vector indicating batches.}

\item{batch2}{optional factor or vector indicating a second series of
batches.}

\item{offset}{The offset is chosen to avoid log2 transform on zeros. Default
to 1.}

\item{...}{Other arguments can be passed to \code{\link[limma]{removeBatchEffect}}.}
}
\value{
An object of argument \code{x} class.
}
\description{
Remove batch effect for copy number: log2 transformed sum of methylation and
unmethylation signals.
}
\details{
Batch effect is removed by calling \code{\link[limma]{removeBatchEffect}}.
  This is only useful when methylation and unmethylation signals are not
  available, such as in objects of \code{\link[minfi]{GenomicRatioSet-class}}.
}
