% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect.R
\name{remove_batch_effect_mum}
\alias{remove_batch_effect_mum}
\title{Remove batch effect: methylation and unmethylation signals.}
\usage{
remove_batch_effect_mum(x, batch = NULL, batch2 = NULL, offset = 1, ...)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}}.}

\item{batch}{factor or vector indicating batches.}

\item{batch2}{optional factor or vector indicating a second series of
batches.}

\item{offset}{The offset is chosen to avoid log2 transform on zeros. Default
to 1.}

\item{...}{Other arguments can be passed to \code{\link[limma]{removeBatchEffect}}.}
}
\value{
An object of argument \code{x} class.
}
\description{
Remove batch effect for methylation and unmethylation signals (log2
transformed) separately. See \href{https://www.ncbi.nlm.nih.gov/pubmed/29539639}{Capper D. DNA methylation-based classification of central nervous system tumours. Nature (2018)}.
}
\details{
Batch effect is removed by calling \code{\link[limma]{removeBatchEffect}}.
  The methylation and unmethylation signals are log2 transformed, removed
  batch effect, and transformed back.
}
