% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{set_annotation_epic_v2}
\alias{set_annotation_epic_v2}
\title{Set annotation for an object of minfi classes of EPIC v2 data.}
\usage{
set_annotation_epic_v2(x)
}
\arguments{
\item{x}{An object of minfi classes.}
}
\value{
An object of minfi classes.
}
\description{
Set annotation for an object of minfi classes of EPIC v2 data.
}
