% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-detection-pvalue.R
\name{summary_detectionP}
\alias{summary_detectionP}
\title{Summarize detection p-values by sample.}
\usage{
summary_detectionP(detP)
}
\arguments{
\item{detP}{a \code{data.frame} returned by \code{\link[minfi]{detectionP}}.}
}
\value{
A \code{data.frame} of summary statistics.
}
\description{
Summarize detection p-values by sample.
}
