% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_qc_metrics.R
\name{write_qc_metrics_excel}
\alias{write_qc_metrics_excel}
\title{Write QC metrics into Excel file.}
\usage{
write_qc_metrics_excel(qc, excel_file, beta_value_distribution_file = NULL)
}
\arguments{
\item{qc}{a \code{data.frame} of QC metrics added to the phenotype data of
the returned \code{\link[minfi]{RGChannelSet-class}} object of
\code{\link{get_qc_metrics}}.}

\item{excel_file}{output Excel file.}

\item{beta_value_distribution_file}{the plot file for beta value distribution.}
}
\value{
a Workbook object.
}
\description{
Write QC metrics into Excel file.
}
