% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-control-probes.R
\name{control_probe_intensities}
\alias{control_probe_intensities}
\title{Get control probes and their raw intensities.}
\usage{
control_probe_intensities(x)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{RGChannelSet-class}}.}
}
\value{
A \code{data.frame} containing the following columns: Address, Type,
  Color, ExtendedType, Channel, Expected_Grn, Expected_Red, Sample, Group,
  Intensity.
}
\description{
Get control probes and their raw intensities.
}
